requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
pypdf==3.17.1
python-docx==1.1.0
python-multipart==0.0.6
openai==1.3.0
tiktoken==0.5.1
numpy==1.24.3
pgvector==0.2.4

